+++ pkgs/dart_mcp/lib/src/api/logging.dart
@@ -26,8 +26,18 @@ extension type SetLevelRequest.fromMap(Map<String, Object?> _value)
   ///
   /// The server should send all logs at this level and higher (i.e., more
   /// severe) to the client as notifications/message.
-  LoggingLevel get level =>
-      LoggingLevel.values.firstWhere((level) => level.name == _value['level']);
+  LoggingLevel get level {
+    final levelName = _value['level'];
+    final foundLevel = LoggingLevel.values.firstWhereOrNull(
+      (level) => level.name == levelName,
+    );
+    if (foundLevel == null) {
+      throw ArgumentError(
+        "Invalid level field in $SetLevelRequest: didn't find level $levelName",
+      );
+    }
+    return foundLevel;
+  }
 }
 
 /// Notification of a log message passed from server to client.