+++ pkgs/dart_mcp/lib/src/api/sampling.dart
@@ -39,8 +39,13 @@ extension type CreateMessageRequest.fromMap(Map<String, Object?> _value)
   });
 
   /// The messages to send to the LLM.
-  List<SamplingMessage> get messages =>
-      (_value['messages'] as List).cast<SamplingMessage>();
+  List<SamplingMessage> get messages {
+    final messages = _value['messages'] as List?;
+    if (messages == null) {
+      throw ArgumentError('Missing messages field in $CreateMessageRequest.');
+    }
+    return messages.cast<SamplingMessage>();
+  }
 
   /// The server's preferences for which model to select.
   ///
@@ -69,7 +74,13 @@ extension type CreateMessageRequest.fromMap(Map<String, Object?> _value)
   /// The maximum number of tokens to sample, as requested by the server.
   ///
   /// The client MAY choose to sample fewer tokens than requested.
-  int get maxTokens => _value['maxTokens'] as int;
+  int get maxTokens {
+    final maxTokens = _value['maxTokens'] as int?;
+    if (maxTokens == null) {
+      throw ArgumentError('Missing maxTokens field in $CreateMessageRequest.');
+    }
+    return maxTokens;
+  }
 
   /// Note: This has no documentation in the specification or schema.
   List<String>? get stopSequences =>